ai-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
